# Performance: Configuration algorithms tuned to minimize the impact of tail effects, now up to 1402 TFLOPS