@LyricZhao 2025-03-10
Tested on our own cluster, it seems a performance fluctuation. And I don't think using less SMs will make performance better and the SM tuning seems not general (e.g. failed to handle `num_sms < 124`). Close it for now.