@sazczmh 2025-03-10
124？Since I don't have the PCIE version of H100, I didn't do the relevant experiments, which can be adjusted as needed, but reducing the number of SMs in a limited range does reduce the impact of the tail effect, and the following are the relevant NCU data。

<img width="1601" alt="image" src="https://github.com/user-attachments/assets/3186522c-fde3-48cc-9313-ca8ee5ed5f9f" />
<img width="1614" alt="image" src="https://github.com/user-attachments/assets/b0bfc4d3-0cde-4b2f-bb5e-bc640af24836" />