@sazczmh 2025-03-10
As far as I know, the get_best_configs function gets the optimal best_util by iterating over block_ms as well as block_ns, which is also intended to minimize the effect of tail effects. But sometimes a **little** reduction in the number of SMs can result in a better best_util, which can slightly improve performance.

For example, 4096x7168x16384 GEMM is quantized into 14 waves, the first 13 waves are very hard to work, but the tasks in the last wave are too easy, and most of the SMs have no work to do. Therefore, a small number of SMs can be removed, so that the tasks assigned to each SM are more balanced.