@LyricZhao 2025-03-10
OK, make sense to me now:

> the first 13 waves are very hard to work

I guess, this should be due to better L2 cache or less possible GPU freq drop?

But I don't think your change is general enough, how about writing a new process (already have the best block sizes, but recompute the minimum num_sms with the same number of waves) at the end of the `get_best_configs`? So that the change will be minimal and supports more GPU archs.

BTW, I think `n_block` naming is confusing, please try to align with the original namings, e.g. `num_sms`/`num_min_sms`.