@sazczmh 2025-03-10
Well, it's likely influenced by multiple factors. To be honest, I'm not entirely clear on the specifics myself. However, empirical test data does show slight performance improvements. Additionally, the current version's power requirements leading to significant frequency reduction should be one of the key optimization directions in the future. The stress-test frequency is only 1.3 GHz, far below the 1.98 GHz baseline.

I'll revise it according to your feedback later.