@sazczmh 2025-03-10
Current configuration sets the minimum value of num_min_sms to num_sms - 8. Performance degradation seems to occur when set lower. The effectiveness of using 112 SMs (Streaming Multiprocessors) remains to be verified through empirical testing.