@sazczmh 2025-03-10
With cluster_size=4 configuration, can we fully utilize all 132 SMs (Streaming Multiprocessors)? This correlates with yield rate – when the yield threshold exceeds 2, it appears to potentially constrain full SM utilization.